moved EdgeWeightsMultiU and NodoeWeightsMultiU to sub-package
