+++ src/dna/metrics/weights/multi/EdgeWeightsMultiU.java
@@ -1,4 +1,4 @@
-package dna.metrics.weights;
+package dna.metrics.weights.multi;
 
 import dna.graph.Graph;
 import dna.graph.IElement;
@@ -9,6 +9,7 @@ import dna.metrics.algorithms.IBeforeEA;
 import dna.metrics.algorithms.IBeforeER;
 import dna.metrics.algorithms.IBeforeEW;
 import dna.metrics.algorithms.IBeforeNR;
+import dna.metrics.weights.EdgeWeights;
 import dna.updates.batch.Batch;
 import dna.updates.update.EdgeAddition;
 import dna.updates.update.EdgeRemoval;
